update: refactor helper/lenet and other classes into their own files
