+++ app.py
@@ -1,19 +1,14 @@
-# app.py
 import base64
 from io import BytesIO
 from random import Random
 import time
 from flask import Flask, flash, json, redirect, render_template, request, url_for
 from flask_socketio import SocketIO, emit
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-from torchvision import datasets, transforms
-import numpy as np
-from skimage.metrics import mean_squared_error, peak_signal_noise_ratio, structural_similarity
 import secrets
 from PIL import Image
-
+from src.dlg import LeNet, ImageProcessing
+from src.utils import Helper, LocalSession
+from torchvision import transforms
 
 app = Flask(__name__)
 app.config.update(
@@ -21,175 +16,9 @@ app.config.update(
 )
 socketio = SocketIO(app)
 
-# Initialize CIFAR and transforms
-dst = datasets.CIFAR100("~/.torch", download=True)
-tp = transforms.Compose([
-    transforms.Resize(32),
-    transforms.CenterCrop(32),
-    transforms.ToTensor()
-])
-tt = transforms.ToPILImage()
-if torch.cuda.is_available():
-    device = "cuda"
-device = "cuda" if torch.cuda.is_available() else "cpu"
-
-class Helper():
-    def __init__(self):
-        pass
-    
-    def label_to_onehot(self, target, num_classes=100):
-        target = torch.unsqueeze(target, 1)
-        onehot_target = torch.zeros(target.size(0), num_classes, device=target.device)
-        onehot_target.scatter_(1, target, 1)
-        return onehot_target
-
-    def cross_entropy_for_onehot(self, pred, target):
-        return torch.mean(torch.sum(- target * F.log_softmax(pred, dim=-1), 1))
-
-    def weights_init(self, m):
-        if hasattr(m, "weight"):
-            m.weight.data.uniform_(-0.5, 0.5)
-        if hasattr(m, "bias"):
-            m.bias.data.uniform_(-0.5, 0.5)
-    
-    def check_similarity(self, ground_truth, reconstructed):
-        ground_truth = np.array(ground_truth)
-        reconstructed = np.array(reconstructed)
-        
-        ground_truth = (ground_truth * 255).astype(np.uint8)
-        reconstructed = (reconstructed * 255).astype(np.uint8)
-        
-        mse = mean_squared_error(ground_truth, reconstructed)
-        psnr = peak_signal_noise_ratio(ground_truth, reconstructed)
-        ssim, _ = structural_similarity(ground_truth, reconstructed, full=True, win_size=3)
-        
-        return mse, psnr, ssim
-    
-    def encode_image(self, image_data):
-        buffer = BytesIO()
-        image_data.save(buffer, format='PNG')
-        return base64.b85encode(buffer.getvalue()).decode()
-    
-    def decode_image(self, b85_string):
-        return base64.b85decode(b85_string)
-    
-class LeNet(nn.Module):
-    def __init__(self, activation_function='sigmoid'):
-        super(LeNet, self).__init__()
-        if activation_function == 'relu':
-            act = nn.ReLU
-        elif activation_function == 'sigmoid':
-            act = nn.Sigmoid
-        elif activation_function == 'tanh':
-            act = nn.Tanh
-        else:
-            act = nn.Sigmoid
-            
-        self.body = nn.Sequential(
-            nn.Conv2d(3, 12, kernel_size=5, padding=5//2, stride=2),
-            act(),
-            nn.Conv2d(12, 12, kernel_size=5, padding=5//2, stride=2),
-            act(),
-            nn.Conv2d(12, 12, kernel_size=5, padding=5//2, stride=1),
-            act(),
-            nn.Conv2d(12, 12, kernel_size=5, padding=5//2, stride=1),
-            act(),
-        )
-        self.fc = nn.Sequential(
-            nn.Linear(768, 100)
-        )
-        
-    def forward(self, x):
-        out = self.body(x)
-        out = out.view(out.size(0), -1)
-        out = self.fc(out)
-        return out
-
-def process_single_image(idx, activation_function):
-    helper = Helper()
-    net = LeNet(activation_function).to(device)
-    net.apply(helper.weights_init)
-    criterion = helper.cross_entropy_for_onehot
-    
-    # Process the image
-    gt_data = tp(dst[idx][0]).to(device)
-    gt_data = gt_data.view(1, *gt_data.size())
-    gt_label = torch.Tensor([dst[idx][1]]).long().to(device)
-    gt_label = gt_label.view(1, )
-    gt_onehot_label = helper.label_to_onehot(gt_label)
-    
-    # Compute original gradient
-    out = net(gt_data)
-    y = criterion(out, gt_onehot_label)
-    dy_dx = torch.autograd.grad(y, net.parameters())
-    original_dy_dx = list((_.detach().clone() for _ in dy_dx))
-    
-    # Generate dummy data and label
-    dummy_data = torch.randn(gt_data.size()).to(device).requires_grad_(True)
-    dummy_label = torch.randn(gt_onehot_label.size()).to(device).requires_grad_(True)
-    optimizer = torch.optim.LBFGS([dummy_data, dummy_label])
-    
-    # Training loop
-    for _ in range(300):
-        def closure():
-            optimizer.zero_grad()
-            pred = net(dummy_data)
-            dummy_onehot_label = F.softmax(dummy_label, dim=-1)
-            dummy_loss = criterion(pred, dummy_onehot_label)
-            dummy_dy_dx = torch.autograd.grad(dummy_loss, net.parameters(), create_graph=True)
-            
-            grad_diff = 0
-            for gx, gy in zip(dummy_dy_dx, original_dy_dx):
-                grad_diff += ((gx - gy) ** 2).sum()
-            grad_diff.backward()
-            
-            return grad_diff
-        
-        optimizer.step(closure)
-    
-    # Get results
-    reconstructed_data = tt(dummy_data[0].cpu())
-    mse, psnr, ssim = helper.check_similarity(tt(gt_data[0].cpu()), reconstructed_data)
-    
-    return {
-        'mse': float(mse),
-        'psnr': float(psnr),
-        'ssim': float(ssim),
-        'image': helper.encode_image(reconstructed_data)
-    }
-
-def process_custom_image(image, activation_function):
-    try:
-        image = image.resize((32,32), Image.LANCZOS)
-        
-        transform = transforms.ToTensor()
-        input_tensor = transform(image).unsqueeze(0)
-        
-        result = process_single_image(0, activation_function)
-        print(result)    
-        return result
-    
-    except Exception as err:
-        print(f"Error processing custom image: {err}")
-    
-    
-class LocalSession():
-    def __init__(self):
-        self.results=[]
-    
-    def add(self, item):
-        self.results.append(item)
-
-    def get_results(self):
-        return self.results
-    
-    def remove(self, item):
-        self.results.remove(item)
-        
-    def clear(self):
-        self.results = []
 
 local_session = LocalSession()
+image_processing = ImageProcessing()
 
 @app.route("/")
 def index():
@@ -213,7 +42,7 @@ def handle_single_process(data):
     cifar_index = data['cifar_index']
     activation_function = data['activation_function']
     try:
-        result = process_single_image(cifar_index, activation_function)
+        result = image_processing.process_single_image(cifar_index, activation_function)
         emit('complete',{'result':result})
     except Exception as err:
         emit('error', str(err))
@@ -233,7 +62,7 @@ def result():
             raise ValueError("No CIFAR index provided")
         
         # Get and encode original image
-        original_img = dst[cifar_index][0]
+        original_img = image_processing.dst[cifar_index][0]
         data['original_image'] = helper.encode_image(original_img)
         
         # Validate reconstructed image exists
@@ -280,7 +109,7 @@ def handle_process(data):
         local_session.clear()
         results = []
         for i, cifar_id in enumerate(range(start_cifar_index, end_cifar_index + 1)):
-            result = process_single_image(cifar_id, activation_function)
+            result = image_processing.process_single_image(cifar_id, activation_function)
             results.append({'cifar_id': cifar_id, **result})
             result_with_id = {'cifar_id': cifar_id, **result}
             local_session.add(result_with_id)
@@ -323,7 +152,7 @@ def chart():
         # Add original images
         for result in results:
             cifar_id = result['cifar_id']
-            original_img = dst[cifar_id][0]
+            original_img = image_processing.dst[cifar_id][0]
             result['original_image'] = helper.encode_image(original_img)
         
         return render_template('chart_multiple.html', results=results)
@@ -403,10 +232,10 @@ def handle_custom_process(data):
         
         # Convert to tensor
         transform = transforms.ToTensor()
-        input_tensor = transform(image).unsqueeze(0).to(device)
+        input_tensor = transform(image).unsqueeze(0).to(image_processing.device)
         
         # Process image using existing pipeline
-        result = process_single_image(0, data['activation_function'])
+        result = image_processing.process_single_image(0, data['activation_function'])
         
         # Add original image to result
         result['original_image'] = data['original_image']
